🐛 fix: useEffect, useRouter 클라이언트 훅을 HeaderAuth로 이동
